+++ lib/axios/interceptor/instance.ts
@@ -1,0 +1,24 @@
+import axios from 'axios';
+import { AxiosInstance } from 'axios/index';
+
+import { setRequestInterceptor } from './request';
+import { setResponseInterceptor } from './response';
+
+//TEST URL
+const BASE_URL = "https://pokeapi.co/api/v2/pokemon/"
+
+const apiClient = axios.create({
+  baseURL: BASE_URL,
+  timeout: 5000,
+  headers: {
+    "Content-Type": "application/json",
+    Accept: "application/json",
+  }
+});
+
+export const initializeApiClient = (): AxiosInstance => {
+  setRequestInterceptor(apiClient);
+  setResponseInterceptor(apiClient);
+  return apiClient;
+}
+export default apiClient;
