+++ src/app/layout.tsx
@@ -3,17 +3,15 @@ import { useEffect } from 'react';
 import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 
-import Initializer from '../../lib/initializer';
-
-import '@/style/globals.css';
+import './globals.css';
 
 const geistSans = localFont({
-  src: '/fonts/GeistVF.woff',
+  src: './fonts/GeistVF.woff',
   variable: '--font-geist-sans',
   weight: '100 900'
 });
 const geistMono = localFont({
-  src: '/fonts/GeistMonoVF.woff',
+  src: './fonts/GeistMonoVF.woff',
   variable: '--font-geist-mono',
   weight: '100 900'
 });
